more fixes for valid build

+++ src/items/itemweapon.h
@@ -6,7 +6,6 @@
 
 #include <commonmacros.h>
 
-
 class ItemWeapon : public Item
 {
 
@@ -17,7 +16,7 @@ public:
     ItemWeapon(ItemWeaponAttribute parameters);
     ~ItemWeapon() override;
 
-    int getHands() override { return _AttributeWeapon.hands;}
+    int getHands() override { return INT(_AttributeWeapon.hands); }
     ItemWeaponDamage getDamage() override;              // Please be sure that you need to set Hand in the inventory
 
 };
